Squashed commit of the following:
Add an argument for the openai service image format

Using LM studio to serve a Qwen Vision LM appeared to crash with base64-encoded png, but not webp. To add compatability with this (and potentially other models) this adds a `openai_image_format` string that gets passed to the mimetype and PIL.save() arguments.

Defaults to ‘webp’ that it was originally hardcoded as.